3d-design-tool/requirements.txt
ollama>=0.3.0
flask>=3.0.0
trimesh>=4.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
scipy>=1.11.0
